multiple: classify_bids errs, get_vars errs

+++ nimb/setup/get_vars.py
@@ -109,14 +109,17 @@ class Get_Vars():
         data['NIMB_PATHS']['miniconda_home']          = new_miniconda_path
         data['NIMB_PATHS']['miniconda_python_run']    = path.join(new_miniconda_path,'bin','python3.7').replace(path.expanduser("~"),"~")
         new_freesurfer_path = get_userdefined_paths('FreeSurfer folder', path.join(NIMB_HOME, '../..', 'freesurfer'), 'freesurfer')
-        FreeSurfer_install = get_yes_no('do you want to install FreeSurfer at the provided location {}? (y/n)'.format(new_freesurfer_path))
-        if FreeSurfer_install == 1:
-            freesurfer_license = get_FS_license()
-        data['FREESURFER']['FreeSurfer_install']      = FreeSurfer_install
+        if not path.exists(new_freesurfer_path):
+            FreeSurfer_install = get_yes_no('do you want to install FreeSurfer at the provided location {}? (y/n)'.format(new_freesurfer_path))
+            data['FREESURFER']['FreeSurfer_install']      = FreeSurfer_install
+            if FreeSurfer_install == 1:
+                freesurfer_license = get_FS_license()
+                data['FREESURFER']['freesurfer_license']  = freesurfer_license
+        else:
+            data['FREESURFER']['FreeSurfer_install']      = 1
         data['FREESURFER']['FREESURFER_HOME']         = new_freesurfer_path
         data['FREESURFER']['FS_SUBJECTS_DIR']         = path.join(new_freesurfer_path, 'subjects')
         data['FREESURFER']['export_FreeSurfer_cmd']   = "export FREESURFER_HOME="+new_freesurfer_path
-        data['FREESURFER']['freesurfer_license']      = freesurfer_license
 
         return data
